Create plugins-update-manager block module
